requirements.txt
openai
fpdf
PyPDF2
streamlit
python-dotenv
reportlab
Pillow
pdf2image
pytesseract